refactor(zero-cache): move table defs to schema/...
